test refactor
I want to see how codecov behaves
