+++ core/src/test/kotlin/com/toasttab/expediter/types/JvmTypeProviderTest.kt
@@ -16,6 +16,7 @@
 package com.toasttab.expediter.types
 
 import com.toasttab.expediter.provider.JvmTypeProvider
+import com.toasttab.model.types.members
 import org.junit.jupiter.api.Test
 import protokt.v1.toasttab.expediter.v1.SymbolicReference
 import strikt.api.expectThat
