+++ model/src/test/kotlin/com/toasttab/expediter/issue/SerializationTest.kt
@@ -1,8 +1,10 @@
 package com.toasttab.expediter.issue
 
-import com.toasttab.expediter.types.MemberAccess
-import com.toasttab.expediter.types.MemberSymbolicReference
-import com.toasttab.expediter.types.MethodAccessType
+import com.toasttab.model.types.MemberAccess
+import com.toasttab.model.types.MemberSymbolicReference
+import com.toasttab.model.types.MethodAccessType
+import com.toasttab.model.issue.Issue
+import com.toasttab.model.issue.IssueReport
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
 import strikt.assertions.containsExactly
